*** empty log message ***
git-svn-id: http://argouml.tigris.org/svn/argouml/trunk@107 a161b567-7d1e-0410-9ef9-912c70fedb3f

+++ src/uci/uml/visual/FigDependency.java
@@ -38,14 +38,15 @@ import uci.gef.*;
 import uci.uml.ui.*;
 import uci.uml.Foundation.Core.*;
 
-public class FigDependency extends FigEdgeLineDotted {
+public class FigDependency extends FigEdgeLine {
 
   public FigDependency(Object edge) {
     super();
     setOwner(edge);
 
     // set whatever arrow heads and colors are appropriate
-    _fig.setLineColor(Color.red);
+    _fig.setLineColor(Color.black);
+    ((FigLine)_fig).setDashed(true);
 
     ArrowHeadGreater endArrow = new ArrowHeadGreater();
     endArrow.setFillColor(Color.red);
@@ -56,6 +57,7 @@ public class FigDependency extends FigEdgeLineDotted {
   public void dispose() {
     if (!(getOwner() instanceof Element)) return;
     Element elmt = (Element) getOwner();
+    if (elmt == null) return;
     Project p = ProjectBrowser.TheInstance.getProject();
     p.moveToTrash(elmt);
     super.dispose();
